requirements.txt
pandas
flask
numpy
seaborn
scikit-learn
xgboost
pymongoMongoClient


#-e. # -e means go and check packages and invoke command find_packages() under current directory(.)